refactor(fiber): move fiber to main dir
