+++ src/internals/CancellationList.ts
@@ -6,9 +6,11 @@ import {ICancellable} from 'ts-scheduler'
 
 export type CancelId = LinkedListNode<ICancellable>
 export class CancellationList implements ICancellable {
+  private cancelled = false
   private readonly Q = new LinkedList<ICancellable>()
 
   public cancel(): void {
+    this.cancelled = true
     let node = this.Q.pop()
     while (node !== undefined) {
       node.cancel()
@@ -16,15 +18,15 @@ export class CancellationList implements ICancellable {
     }
   }
 
-  public cancelById(id: CancelId): void {
-    id.value.cancel()
-    this.remove(id)
-  }
   public push(cancellable: ICancellable): CancelId {
     return this.Q.add(cancellable)
   }
 
   public remove(id: CancelId): void {
     this.Q.remove(id)
+  }
+
+  public isCancelled(): boolean {
+    return this.cancelled
   }
 }
